backend/requirements.txt
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.2
nltk==3.9.1
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==11.0.0
psycopg2-binary==2.9.10
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2024.9.11
six==1.16.0
tqdm==4.66.6
tzdata==2024.2
Unidecode==1.3.8
